+++ contract/src/lib.rs
@@ -27,18 +27,24 @@ const ACCESS_KEY_ALLOWANCE: u128 = 20_000_000_000_000_000_000_000; // 0.02 N
 */ 
 const NEW_ACCOUNT_BASE: u128 = 2_840_000_000_000_000_000_000; // 0.00284 N
 
-const ON_NFT_TRANSFER_GAS: Gas = Gas(20_000_000_000_000); // 20 TGas
-
-const ON_CREATE_ACCOUNT_GAS: Gas = Gas(30_000_000_000_000); // 30 TGas
-const ON_CALLBACK_GAS: Gas = Gas(45_000_000_000_000); // 40 TGas
-
-
+/// Indicates there are no deposit for a callback for better readability.
+const NO_DEPOSIT: u128 = 0;
 
 // Defaulting burnt GAS to be 100 TGas (0.01 $NEAR)
 const BURNT_GAS: u128 = 10_000_000_000_000_000_000_000;
 
-/// Indicates there are no deposit for a callback for better readability.
-const NO_DEPOSIT: u128 = 0;
+/*
+    GAS Constants
+*/
+const GAS_FOR_SIMPLE_NFT_TRANSFER: Gas = Gas(10_000_000_000_000); // 10 TGas
+const GAS_FOR_RESOLVE_TRANSFER: Gas = Gas(5_000_000_000_000 + GAS_FOR_SIMPLE_NFT_TRANSFER.0); //15 TGas
+const GAS_FOR_COMPLEX_NFT_TRANSFER: Gas = Gas(GAS_FOR_SIMPLE_NFT_TRANSFER.0 + GAS_FOR_RESOLVE_TRANSFER.0); //25 TGas
+
+const GAS_FOR_ON_CLAIM: Gas = Gas(30_000_000_000_000 + GAS_FOR_COMPLEX_NFT_TRANSFER.0); // 55 TGas
+const GAS_FOR_CREATE_ACCOUNT: Gas = Gas(30_000_000_000_000); // 30 TGas
+
+const ONE_GIGGA_GAS: u64 = 1_000_000_000;
+
 
 /// Methods callable by the function call access key
 const ACCESS_KEY_METHOD_NAMES: &str = "claim,create_account_and_claim";
@@ -55,8 +61,11 @@ pub struct AccountData {
     pub balance: U128,
     pub token_contract: Option<AccountId>,
     pub nft_id: Option<String>,
+    pub token_sender: Option<AccountId>,
 }
 
+mod claim;
+mod send;
 mod ext_traits;
 mod nft;
 
@@ -89,272 +98,6 @@ impl LinkDropProxy {
         );
 		self.linkdrop_contract = linkdrop_contract;
 	}
-
-	/*
-        user has created a keypair and passes in the public key and attaches some deposit.
-        this will store the account data and allow that key to call claim and create_account_and_claim
-        on this contract.
-
-        The balance is the amount of $NEAR the sender wants the linkdrop to contain.
-    */
-    #[payable]
-    pub fn send(&mut self, public_key: PublicKey, balance: U128) -> Promise {
-        let attached_deposit = env::attached_deposit();
-
-        /*
-            ensure the user attached enough to cover:
-            - storage allowance on the contract for access key and storing account data and pk
-            - access key allowance 
-            - and a balance for the account (which must be greater than new account base)
-        */
-        assert!(
-            attached_deposit >= balance.0 + ACCESS_KEY_ALLOWANCE + STORAGE_ALLOWANCE,
-            "Deposit must be large enough to cover desired balance, access key allowance, and contract storage"
-        );
-
-        assert!(
-            balance.0 >= NEW_ACCOUNT_BASE,
-            "cannot have a desired account balance less than the absolute minimum for creating an account"
-        );
-
-        
-        let pk = public_key;
-        assert!(self.accounts.insert(
-                &pk,
-                &AccountData{
-                    funder_id: env::predecessor_account_id(),
-                    balance: balance,
-                    token_contract: None,
-                    nft_id: None
-                },
-            ).is_none(),
-            "Account for PublicKey exists"
-        );
-
-        // If the user overpaid for the desired linkdrop balance, refund them.
-        if attached_deposit > balance.0 + ACCESS_KEY_ALLOWANCE + STORAGE_ALLOWANCE {
-            Promise::new(env::predecessor_account_id()).transfer(attached_deposit - balance.0 - ACCESS_KEY_ALLOWANCE - STORAGE_ALLOWANCE);
-        }
-
-        /*
-            add the public key as an access key to the contract 
-            which can only call claim and create_account_and_claim on this contract
-        */
-        Promise::new(env::current_account_id()).add_access_key(
-            pk,
-            ACCESS_KEY_ALLOWANCE,
-            env::current_account_id(),
-            ACCESS_KEY_METHOD_NAMES.to_string(),
-        )
-
-        //TODO: do assertions at the end and dynamically calculate required storage? Store this in the account data?
-    }
-
-    /*
-        user has created a bunch of keypairs and passesd in the public keys and attached some deposit.
-        this will store the account data and allow that keys to call claim and create_account_and_claim
-        on this contract.
-
-        The balance is the amount of $NEAR the sender wants each linkdrop to contain.
-    */
-    #[payable]
-    pub fn send_multiple(&mut self, public_keys: Vec<PublicKey>, balance: U128) {
-	    
-        let attached_deposit = env::attached_deposit();
-		let len = public_keys.len() as u128;
-	    
-        /*
-            ensure the user attached enough to cover:
-            - storage allowance on the contract for access key and storing account data and pk
-            - access key allowance 
-            - and a balance for the account (which must be greater than new account base)
-
-            this must be true for every public key passed in.
-        */
-        assert!(
-            attached_deposit >= (balance.0 + ACCESS_KEY_ALLOWANCE + STORAGE_ALLOWANCE) * len,
-            "Deposit must be large enough to cover desired balance, access key allowance, and contract storage for all keys"
-        );
-
-        assert!(
-            balance.0 >= NEW_ACCOUNT_BASE,
-            "cannot have a desired account balance less than the absolute minimum for creating an account"
-        );
-
-		let current_account_id = env::current_account_id();
-		let promise = env::promise_batch_create(&current_account_id);
-        let funder_id = env::predecessor_account_id();
-		
-        // Loop through each public key and insert into the map and create the key
-		for pk in public_keys {
-
-			env::promise_batch_action_add_key_with_function_call(
-				promise, 
-				&pk, 
-				0, 
-				ACCESS_KEY_ALLOWANCE, 
-				&current_account_id, 
-				ACCESS_KEY_METHOD_NAMES
-			);
-			
-            // Ensure none of the public keys exist already
-			assert!(self.accounts.insert(
-                    &pk,
-                    &AccountData{
-                        funder_id: funder_id.clone(),
-                        balance,
-                        token_contract: None,
-                        nft_id: None
-                    },
-                ).is_none(),
-                "Account for PublicKey exists"
-            );
-		}
-
-		env::promise_return(promise);
-    }
-
-	
-    /// Internal method for deleting the used key and removing / returning account data.
-	fn process_claim(&mut self) -> (PublicKey, U128, AccountId, Option<String>, Option<AccountId>) {
-        // Ensure only the current contract is calling the method using the access key
-		assert_eq!(
-            env::predecessor_account_id(),
-            env::current_account_id(),
-            "predecessor != current"
-        );
-
-        // Get the PK of the signer which should be the contract's function call access key
-        let signer_pk = env::signer_account_pk();
-        let AccountData {
-            funder_id,
-            balance,
-            token_contract,
-            nft_id,
-        } = self.accounts
-            .remove(&signer_pk)
-            .expect("Missing public key");
-
-        // Delete the key
-		Promise::new(env::current_account_id()).delete_key(env::signer_account_pk());
-
-        // Return account data info
-		(signer_pk, balance, funder_id, nft_id, token_contract)
-	}
-
-    /// Claim tokens for specific account that are attached to the public key this tx is signed with.
-    pub fn claim(&mut self, account_id: AccountId) -> Promise {
-        // Delete the access key and remove / return account data.
-        let (signer_pk, balance, funder_id, nft_id, token_contract) = self.process_claim();
-		
-        // Send the existing account ID the desired linkdrop balance.
-        Promise::new(account_id.clone()).transfer(balance.0)
-        .then(ext_self::on_claim(
-            account_id,
-			signer_pk,
-            balance,
-            funder_id,
-            nft_id, 
-            token_contract,
-			env::current_account_id(),
-			NO_DEPOSIT,
-			ON_CALLBACK_GAS,
-		))
-    }
-
-    /// Create new account and and claim tokens to it.
-    pub fn create_account_and_claim(
-        &mut self,
-        new_account_id: AccountId,
-        new_public_key: PublicKey,
-    ) -> Promise {
-        let mut used_gas = env::used_gas();
-        let mut prepaid_gas = env::prepaid_gas();
-
-        env::log_str(&format!("Beginning of CAAC used gas: {:?} prepaid gas: {:?}", used_gas, prepaid_gas));
-
-        // Delete the access key and remove / return account data.
-        let (signer_pk, balance, funder_id, nft_id, token_contract) = self.process_claim();
-
-        used_gas = env::used_gas();
-        prepaid_gas = env::prepaid_gas();
-
-        env::log_str(&format!("In CAAC after process claim used gas: {:?} prepaid gas: {:?}", used_gas, prepaid_gas));
-        
-        // CCC to the linkdrop contract to create the account with the desired balance as the linkdrop amount
-        ext_linkdrop::create_account(
-            new_account_id.clone(),
-            new_public_key,
-            self.linkdrop_contract.clone(),
-            balance.0,
-            ON_CREATE_ACCOUNT_GAS,
-        ).then(ext_self::on_claim(
-            new_account_id,
-			signer_pk,
-            balance,
-            funder_id,
-            nft_id, 
-            token_contract,
-			env::current_account_id(),
-			NO_DEPOSIT,
-			ON_CALLBACK_GAS,
-		))
-    }
-
-	/// self callback checks if account was created successfully or not. If yes, refunds excess storage, sends NFTs, FTs etc..
-    pub fn on_claim(&mut self, new_account_id: AccountId, pk: PublicKey, balance: U128, funder_id: AccountId, nft_id: Option<String>, token_contract: Option<AccountId>) -> bool {
-        let mut used_gas = env::used_gas();
-        let mut prepaid_gas = env::prepaid_gas();
-
-        env::log_str(&format!("Beginning of on claim used gas: {:?} prepaid gas: {:?}", used_gas, prepaid_gas));
-
-        assert_eq!(
-            env::predecessor_account_id(),
-            env::current_account_id(),
-            "predecessor != current"
-        );
-		assert_eq!(env::promise_results_count(), 1, "no promise result");
-        let creation_succeeded = matches!(env::promise_result(0), PromiseResult::Successful(_));
-
-        // Default amount to refund to be everything except balance and burnt GAS since balance was sent to new account.
-        let mut amount_to_refund =  ACCESS_KEY_ALLOWANCE + STORAGE_ALLOWANCE - BURNT_GAS;
-        
-        // If not successful, the balance is added to the amount to refund since it was never transferred.
-        if !creation_succeeded {
-            amount_to_refund += balance.0
-        }
-
-        used_gas = env::used_gas();
-        prepaid_gas = env::prepaid_gas();
-
-        env::log_str(&format!("In on claim before refund used gas: {:?} prepaid gas: {:?}", used_gas, prepaid_gas));
-
-        Promise::new(funder_id).transfer(amount_to_refund);
-
-        // If NFT isn't successfully transferred, should send it back to it's original owner? 
-        if nft_id.is_some() {
-            let nft_contract_id = token_contract.expect("no contract ID found");
-            let token_id = nft_id.expect("no token Id found");
-
-            used_gas = env::used_gas();
-            prepaid_gas = env::prepaid_gas();
-
-            env::log_str(&format!("In on claim before nft transfer used gas: {:?} prepaid gas: {:?}", used_gas, prepaid_gas));
-
-             // CCC to the NFT contract to transfer the token.
-             ext_nft_contract::nft_transfer(
-                new_account_id, 
-                token_id,
-                None,
-                Some("Linkdroped NFT".to_string()),
-                nft_contract_id,
-                1,
-                ON_NFT_TRANSFER_GAS,
-            );
-        }
-
-        creation_succeeded
-    }
 
     /// Returns the balance associated with given key. This is used by the NEAR wallet to display the amount of the linkdrop
     pub fn get_key_balance(&self, key: PublicKey) -> U128 {
