+++ neardev/dev-account.env
@@ -1,1 +1,1 @@
-CONTRACT_NAME=dev-1652378734203-20540771367880
+CONTRACT_NAME=dev-1652394165698-42979226360879
